+++ lib/queue_bus/driver.rb
@@ -21,10 +21,22 @@ module QueueBus
       def perform(attributes = {})
         raise 'No attributes passed' if attributes.empty?
 
-        ::QueueBus.log_worker("Driver running: #{attributes.inspect}")
+        log_metadata = {
+          event_type: attributes['bus_event_type'],
+          bus_id: attributes['bus_id'],
+          attributes_count: attributes.keys.size
+        }
+        ::QueueBus.log_worker('Driver processing event', log_metadata)
 
         subscription_matches(attributes).each do |sub|
-          ::QueueBus.log_worker("  ...sending to #{sub.queue_name} queue with class #{sub.class_name} for app #{sub.app_key} because of subscription: #{sub.key}")
+          sub_metadata = {
+            queue_name: sub.queue_name,
+            class_name: sub.class_name,
+            app_key: sub.app_key,
+            subscription_key: sub.key,
+            event_type: attributes['bus_event_type']
+          }
+          ::QueueBus.log_worker('Dispatching to subscriber', sub_metadata)
 
           bus_attr = {  'bus_driven_at' => Time.now.to_i,
                         'bus_rider_queue' => sub.queue_name,